Improve rendering of home, content and news pages
Bring the various pages into a common structure of:

- pyconau2017/content_page.html which extends site_base_wagtail which extends site_base and defines a single content block called "page_content" which is enclosed in a single div class "page-content". This div's styling (pyconau.css) is the same as the jumbotron, but I don't like use continuing to enclose everything in jumbotron, it feels wrong :-)
- the various cms pages (content_page, news_page, etc.) all extend abstract_content_page which extends the above pyconau2017/content_page. The abstract page just adds a bunch of wagtail stuff, and leaves the page_content block alone for the underlying content_page, news_page etc. to fill in. You'll see that those pages are quite simple now.

Note that the home_page type is *raw HTML* and I guess we'll just deal with that.

There was a *lot* of duplication in the wagtail cms_pages templates (like multiple levels of setting page.title into the head_title block). I've removed those too.

Also fixed structure in public-facing and admin-side schedule pages.