A small change to address a few pages that were aborting due to attempt to use (non-existent, in py3) unicode method.
Commented out stuff in settings.py that disbled parts of the django-debug-toolbar.
Turned out this wasn't necessary as all of the fixes required are in wagtail.

See https://github.com/wagtail/wagtail/pull/3738/files for those fixes.
(These should have made it into wagtail main branch by now, but, as of
this commit they haven't yet for some reason.  So you'll probably need
to monkey-patch wagtail to get this to work.  ~N6)